try to fix edge runtime error cf workers
